requirements.txt
pandas>=2.1
numpy>=1.26,<2
streamlit>=1.35
plotly>=5.20
pyyaml>=6.0
openpyxl>=3.1
numexpr>=2.9
pytest>=8.0
